+++ src/main/java/com/brotes/api/modelo/itemPedidoRecurrente/DatosRegistroItemPedidoRecurrente.java
@@ -1,3 +1,3 @@
 package com.brotes.api.modelo.itemPedidoRecurrente;
 
-public record DatosRegistroItemPedidoRecurrente(int cantidad, Long idPedido, Long idProducto) {}
+public record DatosRegistroItemPedidoRecurrente(Double cantidad, Long idPedido, Long idProducto) {}
