+++ src/main/java/com/brotes/api/modelo/pedidos/ItemPlanillaProduccion.java
@@ -1,4 +1,4 @@
 package com.brotes.api.modelo.pedidos;
 
-public record ItemPlanillaProduccion(String nombre, Integer cantidad) {
+public record ItemPlanillaProduccion(String nombre, Double cantidad) {
 }
