+++ src/main/java/com/brotes/api/modelo/producto/DatosProductoPedido.java
@@ -6,5 +6,5 @@ public record DatosProductoPedido(
         @NotNull
         Long id,
         @NotNull
-        int cantidad) {
+        Double cantidad) {
 }
